--- app/(request)/public-cloud/create/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div>CREATE REQUEST</div>;
-}
