--- app/(request)/public-cloud/edit/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div>EDIT REQUEST</div>;
-}
