+++ app/api/private-cloud/delete/[id]/route.ts
