+++ app/api/public-cloud/project/[licencePlate]/route.ts
