+++ components/form/AsyncAutocomplete.tsx
@@ -60,7 +60,6 @@ export default function AsyncAutocomplete({
   } = useFormContext();
 
   const email = watch(name + ".email");
-  console.log("EMAIL", email);
 
   const {
     data: people,
